pyproject.toml
[tool.poetry]
name = "parking-cecar-api-v2"
version = "0.1.0"
description = "V2 of the license plate detector api, extended with celery-redis message broker for background task processing"
authors = ["ReadingShades <42413498+ReadingShades@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests-file = "^1.5.1"
torch = "^2.0.1"
transformers = "^4.33.2"
validators = "^0.22.0"
easyocr = "^1.7.1"
opencv-python = "^4.8.0.76"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
redis = "^5.0.0"
celery = "^5.3.4"
gevent = "^23.9.1"
gunicorn = "^21.2.0"
uvicorn = "^0.23.2"
drf-spectacular = "^0.26.5"
django-cors-headers = "^4.2.0"
setuptools = "^68.2.2"
pip-tools = "^7.3.0"
cryptography = "^41.0.4"
ultralytics = "^8.0.223"

[tool.poetry.group.dev.dependencies]
mypy = "^1.5.1"
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
pytest-django = "^4.5.2"
nox = "^2023.4.22"
isort = "^5.12.0"
black = "^23.9.1"
flake8 = "^6.1.0"
ruff = "^0.0.291"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 80
target-version = ["py38"]
exclude = '''
/(
    \.git
  | \.venv
  | \.direnv
  | _build
  | build
  | dist
  | \.env
  | \.vscode
  | \.cache
  | github
  | __pycache__
  | .cloudignore
)/
'''

[tool.isort]
py_version = "38"
line_length = 80
lines_after_imports = 2
lines_between_sections = 1
multi_line_output = 3
force_grid_wrap = 0
include_trailing_comma = "true"
combine_as_imports = "true"
use_parentheses = "true"
skip = ["./docs", ".direnv"]
skip_glob = ["*/__init__.py"]
skip_gitignore = "true"
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
default_section = "THIRDPARTY"

[tool.mypy]
python_version = 3.11
warn_unused_configs = true
ignore_missing_imports = true